Refactor: extract backoff
