+++ src/sinks/util/retries.rs
@@ -190,86 +190,6 @@ impl RetryAction {
     }
 }
 
-// `tokio-retry` crate
-// MIT License
-// Copyright (c) 2017 Sam Rijs
-//
-/// A retry strategy driven by exponential back-off.
-///
-/// The power corresponds to the number of past attempts.
-#[derive(Debug, Clone)]
-pub struct ExponentialBackoff {
-    current: u64,
-    base: u64,
-    factor: u64,
-    max_delay: Option<Duration>,
-}
-
-impl ExponentialBackoff {
-    /// Constructs a new exponential back-off strategy,
-    /// given a base duration in milliseconds.
-    ///
-    /// The resulting duration is calculated by taking the base to the `n`-th power,
-    /// where `n` denotes the number of past attempts.
-    pub const fn from_millis(base: u64) -> ExponentialBackoff {
-        ExponentialBackoff {
-            current: base,
-            base,
-            factor: 1u64,
-            max_delay: None,
-        }
-    }
-
-    /// A multiplicative factor that will be applied to the retry delay.
-    ///
-    /// For example, using a factor of `1000` will make each delay in units of seconds.
-    ///
-    /// Default factor is `1`.
-    pub const fn factor(mut self, factor: u64) -> ExponentialBackoff {
-        self.factor = factor;
-        self
-    }
-
-    /// Apply a maximum delay. No retry delay will be longer than this `Duration`.
-    pub const fn max_delay(mut self, duration: Duration) -> ExponentialBackoff {
-        self.max_delay = Some(duration);
-        self
-    }
-
-    /// Resents the exponential back-off strategy to its initial state.
-    pub fn reset(&mut self) {
-        self.current = self.base;
-    }
-}
-
-impl Iterator for ExponentialBackoff {
-    type Item = Duration;
-
-    fn next(&mut self) -> Option<Duration> {
-        // set delay duration by applying factor
-        let duration = if let Some(duration) = self.current.checked_mul(self.factor) {
-            Duration::from_millis(duration)
-        } else {
-            Duration::from_millis(std::u64::MAX)
-        };
-
-        // check if we reached max delay
-        if let Some(ref max_delay) = self.max_delay {
-            if duration > *max_delay {
-                return Some(*max_delay);
-            }
-        }
-
-        if let Some(next) = self.current.checked_mul(self.base) {
-            self.current = next;
-        } else {
-            self.current = std::u64::MAX;
-        }
-
-        Some(duration)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::{fmt, time::Duration};
