+++ src/sinks/util/service/net/mod.rs
@@ -15,10 +15,11 @@ use std::{
 use std::path::PathBuf;
 
 use crate::{
+    common::backoff::ExponentialBackoff,
     internal_events::{
         SocketOutgoingConnectionError, TcpSocketConnectionEstablished, UdpSendIncompleteError,
     },
-    sinks::{util::retries::ExponentialBackoff, Healthcheck},
+    sinks::Healthcheck,
 };
 
 #[cfg(unix)]
