+++ src/sinks/util/tcp.rs
@@ -23,6 +23,7 @@ use vector_core::{ByteSizeOf, EstimatedJsonEncodedSizeOf};
 
 use crate::{
     codecs::Transformer,
+    common::backoff::ExponentialBackoff,
     dns,
     event::Event,
     internal_events::{
@@ -31,7 +32,6 @@ use crate::{
     },
     sinks::{
         util::{
-            retries::ExponentialBackoff,
             socket_bytes_sink::{BytesSink, ShutdownCheck},
             EncodedEvent, SinkBuildError, StreamSink,
         },
